Many Solutions of BOJ 2225

+++ Baekjoon/C++/Baekjoon02225_SumDecomposition.cpp
@@ -10,11 +10,11 @@ int main(void) {
 	int N, K;
 	cin >> N >> K;
 
-	long long dp[201][201] = { 0LL, };
+	int dp[201][201] = { 0, };
 	for ( int i = 1; i <= K; i++ )
 		dp[0][i] = 1;
 
-	// Method 1: Sum of All
+	// Solution 1: Sum of All
 	for ( int i = 1; i <= N; i++ ) {
 		dp[i][1] = 1;
 		for ( int j = 2; j <= K; j++ )
@@ -24,7 +24,7 @@ int main(void) {
 
 	cout << dp[N][K] << '\n';
 
-	// Method 2: Using prefix sum
+	// Solution 2: Using prefix sum
 	for ( int i = 1; i <= N; i++ ) {
 		dp[i][1] = dp[i - 1][1] + 1;
 		for ( int j = 2; j <= K; j++ )
@@ -33,7 +33,7 @@ int main(void) {
 
 	cout << (dp[N][K] - dp[N - 1][K] + MOD) % MOD << '\n';
 
-	// Method 3: Using value caculated already
+	// Solution 3: Using value caculated already
 	for ( int i = 1; i <= N; i++ ) {
 		dp[i][1] = 1;
 		for ( int j = 2; j <= K; j++ )
